V0.0.8
Added some more unit tests, refactorings.  The Product class is slowly
becoming less and less of the Dumb Data Object I originally intended it
to be.  I was considering changing it from the strangely implemented
multiton it is now to a complex enum, but now it's moving in the
direction of a full java class.  I might still change it into a full
multiton though.